change from double to float since don't need that many bits to represent the numbers in the simulation

+++ include/axis.hpp
@@ -5,13 +5,13 @@
 #include <memory>
 
 class Hole {
-    double position_;
-    double width_;
-    double start_pos_, end_pos_;
-    double influxProbability_;
-    double efluxProbability_;
+    float position_;
+    float width_;
+    float start_pos_, end_pos_;
+    float influxProbability_;
+    float efluxProbability_;
 public:
-    Hole(double pos, double width, double influxProb, double efluxProb);
+    Hole(float pos, float width, float influxProb, float efluxProb);
     // check if a given particle is inside this hole
     bool isParticleInside(const Particle& p); 
     float getCenterPosition() const;
@@ -25,18 +25,18 @@ public:
 
 
 class Axis {
-    double length_;
-    double start_position_;
-    double end_position_;
+    float length_;
+    float start_position_;
+    float end_position_;
     std::vector<Hole> hole_list_;
     std::vector<std::shared_ptr<Particle>> particle_list_;
 public:
-    Axis(double length);
+    Axis(float length);
     void addHole(const Hole& hole);
     void addParticle(const Particle& particle);
     void rmParticle(const Particle& particle);
     // move all particles along the axis
-    void moveParticles(double dt); 
+    void moveParticles(float dt); 
     // handle influx/eflux of particles in each hole
     void handleHoles(); 
 };
